+++ pyproject.toml
@@ -26,11 +26,7 @@ include = [
 
 [project.optional-dependencies]
 all = [
-    "redis>=6.2.0",
-    "firebase_admin>=6.4.0",
     "python_core @ git+https://bar-fingerman-dev:fc9f605912b623be07ebf0eaf32cd83671833a82@github.com/Bria-AI/python_core.git@1.4",
-    "boto3>=1.34.0",
-    "slack_sdk==3.26.1",
     "httpx==0.28.1",
     "pydantic==2.11.10",
     "python_dotenv==1.1.1",
@@ -45,7 +41,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/bria_internal"]
+packages = ["src/"]
 
 
 [tool.ruff.lint]
@@ -66,3 +62,10 @@ dev = [
 [tool.pytest.ini_options]
 log_cli = true
 python_files = "**/*_tests.py"
+
+[tool.uv]
+dev-dependencies = [
+    "pytest>=8.4.0",
+    "pytest-mock>=3.14.1", 
+    "ruff>=0.12.0",
+]
