+++ src/bria_engine_api/apis/image_editing/mask_based_editing.py
@@ -1,15 +1,15 @@
 from httpx import Response
 
-from bria_engine_api.apis.status.status import StatusAPI
+from bria_engine_api.apis.status import StatusAPI
 from bria_engine_api.constants import BriaEngineAPIRoutes
 from bria_engine_api.enable_sync_decorator import enable_run_synchronously
 from bria_engine_api.engine_client import BriaEngineClient
 from bria_engine_api.exceptions.engine_api_exception import EngineAPIException
-from bria_engine_api.schemas.image_editing_apis.canvas_editing import GetMasksRequestPayload, ObjectEraserRequestPayload, ObjectGenFillRequestPayload
+from bria_engine_api.schemas.image_editing_apis import GetMasksRequestPayload, ObjectEraserRequestPayload, ObjectGenFillRequestPayload
 from bria_engine_api.schemas.status_api import StatusAPIResponse
 
 
-class CanvasEditingAPI:
+class MasksBasedEditingAPI:
     def __init__(self, engine_client: BriaEngineClient, status_api: StatusAPI):
         self.__engine_client = engine_client
         self.__status_api = status_api
