+++ src/bria_engine_api/schemas/image_editing_apis/__init__.py
@@ -1,17 +1,17 @@
+from bria_engine_api.schemas.base_models import APIPayloadModel, ContentModeratedPayloadModel, PromptContentModeratedPayloadModel
 from bria_engine_api.schemas.image_editing_apis.background_editing import (
     BlurBackgroundRequestPayload,
     RemoveBackgroundRequestPayload,
     ReplaceBackgroundMode,
     ReplaceBackgroundRequestPayload,
 )
-from bria_engine_api.schemas.image_editing_apis.canvas_editing import (
-    CanvasOperationMaskType,
+from bria_engine_api.schemas.image_editing_apis.foreground_editing import CropOutRequestPayload, ReplaceForegroundRequestPayload
+from bria_engine_api.schemas.image_editing_apis.mask_based_editing import (
     GetMasksRequestPayload,
+    MaskType,
     ObjectEraserRequestPayload,
     ObjectGenFillRequestPayload,
 )
-from bria_engine_api.schemas.image_editing_apis.foreground_editing import CropOutRequestPayload, ReplaceForegroundRequestPayload
-from bria_engine_api.schemas.image_editing_apis.generic_payloads import APIPayloadModel, ContentModeratedPayloadModel, PromptContentModeratedPayloadModel
 from bria_engine_api.schemas.image_editing_apis.size_editing import (
     EnhanceImageRequestPayload,
     ExpandImageRequestPayload,
@@ -27,7 +27,7 @@ __all__ = [
     "ReplaceBackgroundMode",
     "ReplaceBackgroundRequestPayload",
     "BlurBackgroundRequestPayload",
-    "CanvasOperationMaskType",
+    "MaskType",
     "ObjectEraserRequestPayload",
     "ObjectGenFillRequestPayload",
     "GetMasksRequestPayload",
