+++ src/bria_sdk.py
@@ -1,12 +1,8 @@
 from contextvars import ContextVar
 
-from .bria_engine_api.apis.image_editing import ImageEditingAPI
-from .bria_engine_api.apis.status.status import StatusAPI
-from .bria_engine_api.engine_client import BriaEngineClient
+from bria_engine_api.apis import EngineAPIs
 
 
 class BriaSDK:
     def __init__(self, api_token_ctx: ContextVar[str] | None = None, jwt_token_ctx: ContextVar[str] | None = None):
-        self.__client = BriaEngineClient(api_token_ctx, jwt_token_ctx)
-        self.status = StatusAPI(self.__client)
-        self.image_editing = ImageEditingAPI(self.__client, self.status)
+        self.engine_apis = EngineAPIs(api_token_ctx, jwt_token_ctx)
